Fix Claude peak hours lint
